sos_guardian_final/sos_final/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
pydantic>=2.9.2
httpx>=0.27.0
